# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vaishnavi Residency Hotel Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /guests:
    get:
      operationId: getGuests
      tags: [guests]
      summary: List all guests
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create a new guest entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Guest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}/checkin:
    post:
      operationId: checkInGuest
      tags: [guests]
      summary: Check in a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}/checkout:
    post:
      operationId: checkOutGuest
      tags: [guests]
      summary: Check out a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: guestId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, role, name]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, receptionist]
        createdAt:
          type: string
...
    Guest:
      type: object
      required: [id, billNumber, guestName, roomId, checkInDate, status]
      properties:
        id:
          type: integer
        billNumber:
          type: string
        date:
          type: string
        time:
          type: string
        guestName:
          type: string
        gender:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        mobile:
          type: string
        alternativeMobile:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        aadhaarNumber:
          type: ["string", "null"]
        idProofType:
          type: ["string", "null"]
        pax:
          type: integer
        roomId:
          type: integer
        roomNumber:
          type: ["string", "null"]
        checkInDate:
          type: string
        checkInTime:
          type: ["string", "null"]
        expectedCheckOutDate:
          type: ["string", "null"]
        expectedCheckOutTime:
          type: ["string", "null"]
        actualCheckOutDate:
          type: ["string", "null"]
        actualCheckOutTime:
          type: ["string", "null"]
        numberOfDays:
          type: ["integer", "null"]
        rentPerDay:
          type: number
        extraCharges:
          type: number
        discount:
          type: number
        gst:
          type: number
        totalAmount:
          type: number
        advancePaid:
          type: number
        balanceAmount:
          type: number
        paymentMode:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        receptionistName:
          type: ["string", "null"]
        status:
          type: string
          enum: [booked, checked-in, checked-out, cancelled]
        createdAt:
          type: string
    GuestInput:
      type: object
      required: [guestName, mobile, roomId, checkInDate, rentPerDay]
      properties:
        billNumber:
          type: string
        date:
          type: string
        time:
          type: string
        guestName:
          type: string
        gender:
          type: string
        age:
          type: integer
        mobile:
          type: string
        alternativeMobile:
          type: string
        address:
          type: string
        aadhaarNumber:
          type: string
        idProofType:
          type: string
        pax:
          type: integer
        roomId:
          type: integer
        checkInDate:
          type: string
        checkInTime:
          type: string
        expectedCheckOutDate:
          type: string
        expectedCheckOutTime:
          type: string
        numberOfDays:
          type: integer
        rentPerDay:
          type: number
        extraCharges:
          type: number
        discount:
          type: number
        gst:
          type: number
        totalAmount:
          type: number
        advancePaid:
          type: number
        balanceAmount:
          type: number
        paymentMode:
          type: string
        remarks:
          type: string
        receptionistName:
          type: string
    GuestUpdate:
      type: object
      properties:
        guestName:
          type: string
        gender:
          type: string
        age:
          type: integer
        mobile:
          type: string
        alternativeMobile:
          type: string
        address:
          type: string
        aadhaarNumber:
          type: string
        idProofType:
          type: string
        pax:
          type: integer
        roomId:
          type: integer
        checkInDate:
          type: string
        checkInTime:
          type: string
        expectedCheckOutDate:
          type: string
        expectedCheckOutTime:
          type: string
        numberOfDays:
          type: integer
        rentPerDay:
          type: number
        extraCharges:
          type: number
        discount:
          type: number
        gst:
          type: number
        totalAmount:
          type: number
        advancePaid:
          type: number
        balanceAmount:
          type: number
        paymentMode:
          type: string
        remarks:
          type: string
        receptionistName:
          type: string
        status:
          type: string
          enum: [booked, checked-in, checked-out, cancelled]
    CheckoutInput:
      type: object
      properties:
        actualCheckOutDate:
          type: string
        actualCheckOutTime:
          type: string
        finalAmount:
          type: number
        balanceAmount:
          type: number
        paymentMode:
          type: string
...
    Payment:
      type: object
      required: [id, voucherNumber, type, guestId, amount, paymentMode]
      properties:
        id:
          type: integer
        voucherNumber:
          type: string
        type:
          type: string
          enum: [receipt, payment]
        guestId:
          type: integer
        guestName:
          type: ["string", "null"]
        billNumber:
          type: ["string", "null"]
        amount:
          type: number
        paymentMode:
          type: string
        remarks:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [type, guestId, amount, paymentMode]
      properties:
        type:
          type: string
          enum: [receipt, payment]
        guestId:
          type: integer
        amount:
          type: number
        paymentMode:
          type: string
        date:
          type: string
        remarks:
          type: string
...