# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vaishnavi Residency Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestUpdateStatus } from './guestUpdateStatus';

export interface GuestUpdate {
  guestName?: string;
  gender?: string;
  age?: number;
  mobile?: string;
  alternativeMobile?: string;
  address?: string;
  aadhaarNumber?: string;
  idProofType?: string;
  pax?: number;
  roomId?: number;
  checkInDate?: string;
  checkInTime?: string;
  expectedCheckOutDate?: string;
  expectedCheckOutTime?: string;
  numberOfDays?: number;
  rentPerDay?: number;
  extraCharges?: number;
  discount?: number;
  gst?: number;
  totalAmount?: number;
  advancePaid?: number;
  balanceAmount?: number;
  paymentMode?: string;
  remarks?: string;
  receptionistName?: string;
  status?: GuestUpdateStatus;
}